+++ src/OpenTelemetry.Instrumentation.Runtime/RuntimeMetricsOptions.cs
@@ -46,6 +46,11 @@ namespace OpenTelemetry.Instrumentation.Runtime
         public bool? AssembliesEnabled { get; set; }
 
         /// <summary>
+        /// Gets or sets a value indicating whether exception count metrics should be collected.
+        /// </summary>
+        public bool? ExceptionCountEnabled { get; set; }
+
+        /// <summary>
         /// Gets a value indicating whether all metrics are enabled.
         /// </summary>
         internal bool IsAllEnabled => this.GcEnabled == null
@@ -55,7 +60,8 @@ namespace OpenTelemetry.Instrumentation.Runtime
 #if NETCOREAPP3_1_OR_GREATER
         && this.ThreadingEnabled == null
 #endif
-        && this.AssembliesEnabled == null;
+        && this.AssembliesEnabled == null
+        && this.ExceptionCountEnabled == null;
 
         /// <summary>
         /// Gets a value indicating whether garbage collection metrics is enabled.
@@ -80,5 +86,10 @@ namespace OpenTelemetry.Instrumentation.Runtime
         /// Gets a value indicating whether assembly metrics is enabled.
         /// </summary>
         internal bool IsAssembliesEnabled => this.AssembliesEnabled == true || this.IsAllEnabled;
+
+        /// <summary>
+        /// Gets a value indicating whether exception count metrics is enabled.
+        /// </summary>
+        internal bool IsExceptionCountEnabled => this.ExceptionCountEnabled == true || this.IsAllEnabled;
     }
 }